RF+BIG: Metadata "Parser" -> "Extractor"
That is the way people talk about it, and we have plenty of other
things that are called parser and do not extract metadata.

At the same time, this commit also introduces consistent logger
names from all parsers.
