+++ datalad/metadata/extractors/tests/test_image.py
@@ -6,13 +6,13 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""Test NIfTI parser"""
+"""Test image extractor"""
 
 from datalad.tests.utils import SkipTest
 try:
-    import nibabel
+    import PIL
 except ImportError:
-    raise SkipTest
+    raise SkipTest("No PIL module available")
 
 from shutil import copy
 from os.path import dirname
@@ -27,49 +27,30 @@ from datalad.tests.utils import assert_in
 
 
 target = {
-    "description": "FSL5.0",
-    "spatial_resolution(mm)": [2.0, 2.0, 2.0],
-    "temporal_spacing(s)": 6.0,
-    "datatype": "int16",
-    "dim": [4, 91, 109, 91, 2, 1, 1, 1],
-    "pixdim": [-1.0, 2.0, 2.0, 2.0, 6.0, 1.0, 1.0, 1.0],
-    "xyz_unit": "millimiter (uo:0000016)",
-    "t_unit": "second (uo:0000010)",
-    "cal_min": 3000.0,
-    "cal_max": 8000.0,
-    "toffset": 0.0,
-    "vox_offset": 0.0,
-    "intent": "none",
-    "sizeof_hdr": 348,
-    "magic": "n+1",
-    "sform_code": "mni",
-    "qform_code": "mni",
-    "freq_axis": None,
-    "phase_axis": None,
-    "slice_axis": None,
-    "slice_start": 0,
-    "slice_duration": 0.0,
-    "slice_order": "unknown",
-    "slice_end": 0,
+    "dcterms:SizeOrDuration": [4, 3],
+    "color_mode": "3x8-bit pixels, true color",
+    "type": "dctype:Image",
+    "spatial_resolution(dpi)": [72, 72],
+    "format": "JPEG (ISO 10918)"
 }
 
 
 @with_tempfile(mkdir=True)
-def test_nifti(path):
+def test_image(path):
     ds = Dataset(path).create()
-    ds.config.add('datalad.metadata.nativetype', 'nifti1', where='dataset')
+    ds.config.add('datalad.metadata.nativetype', 'image', where='dataset')
     copy(
-        opj(dirname(dirname(dirname(__file__))), 'tests', 'data', 'nifti1.nii.gz'),
+        opj(dirname(dirname(dirname(__file__))), 'tests', 'data', 'exif.jpg'),
         path)
     ds.add('.')
     ok_clean_git(ds.path)
     res = ds.aggregate_metadata()
     assert_status('ok', res)
-    res = ds.metadata('nifti1.nii.gz')
+    res = ds.metadata('exif.jpg')
     assert_result_count(res, 1)
 
-    # from this parser
-    meta = res[0]['metadata']['nifti1']
+    # from this extractor
+    meta = res[0]['metadata']['image']
     for k, v in target.items():
         eq_(meta[k], v)
