+++ datalad/metadata/extractors/tests/test_nifti1.py
@@ -6,11 +6,11 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""Test audio parser"""
+"""Test NIfTI extractor"""
 
 from datalad.tests.utils import SkipTest
 try:
-    import mutagen
+    import nibabel
 except ImportError:
     raise SkipTest
 
@@ -27,35 +27,49 @@ from datalad.tests.utils import assert_in
 
 
 target = {
-    "format": "mime:audio/mp3",
-    "duration(s)": 1.0,
-    "name": "dltracktitle",
-    "music:album": "dlalbumtitle",
-    "music:artist": "dlartist",
-    "music:channels": 1,
-    "music:sample_rate": 44100,
-    "music:Genre": "dlgenre",
-    "date": "",
-    "tracknumber": "dltracknumber",
+    "description": "FSL5.0",
+    "spatial_resolution(mm)": [2.0, 2.0, 2.0],
+    "temporal_spacing(s)": 6.0,
+    "datatype": "int16",
+    "dim": [4, 91, 109, 91, 2, 1, 1, 1],
+    "pixdim": [-1.0, 2.0, 2.0, 2.0, 6.0, 1.0, 1.0, 1.0],
+    "xyz_unit": "millimiter (uo:0000016)",
+    "t_unit": "second (uo:0000010)",
+    "cal_min": 3000.0,
+    "cal_max": 8000.0,
+    "toffset": 0.0,
+    "vox_offset": 0.0,
+    "intent": "none",
+    "sizeof_hdr": 348,
+    "magic": "n+1",
+    "sform_code": "mni",
+    "qform_code": "mni",
+    "freq_axis": None,
+    "phase_axis": None,
+    "slice_axis": None,
+    "slice_start": 0,
+    "slice_duration": 0.0,
+    "slice_order": "unknown",
+    "slice_end": 0,
 }
 
 
 @with_tempfile(mkdir=True)
-def test_audio(path):
+def test_nifti(path):
     ds = Dataset(path).create()
-    ds.config.add('datalad.metadata.nativetype', 'audio', where='dataset')
+    ds.config.add('datalad.metadata.nativetype', 'nifti1', where='dataset')
     copy(
-        opj(dirname(dirname(dirname(__file__))), 'tests', 'data', 'audio.mp3'),
+        opj(dirname(dirname(dirname(__file__))), 'tests', 'data', 'nifti1.nii.gz'),
         path)
     ds.add('.')
     ok_clean_git(ds.path)
     res = ds.aggregate_metadata()
     assert_status('ok', res)
-    res = ds.metadata('audio.mp3')
+    res = ds.metadata('nifti1.nii.gz')
     assert_result_count(res, 1)
 
-    # from this parser
-    meta = res[0]['metadata']['audio']
+    # from this extractor
+    meta = res[0]['metadata']['nifti1']
     for k, v in target.items():
         eq_(meta[k], v)
