+++ datalad/metadata/extractors/tests/test_xmp.py
@@ -6,13 +6,13 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""Test image parser"""
+"""Test XMP extractor"""
 
 from datalad.tests.utils import SkipTest
 try:
-    import PIL
+    import libxmp
 except ImportError:
-    raise SkipTest("No PIL module available")
+    raise SkipTest
 
 from shutil import copy
 from os.path import dirname
@@ -27,30 +27,36 @@ from datalad.tests.utils import assert_in
 
 
 target = {
-    "dcterms:SizeOrDuration": [4, 3],
-    "color_mode": "3x8-bit pixels, true color",
-    "type": "dctype:Image",
-    "spatial_resolution(dpi)": [72, 72],
-    "format": "JPEG (ISO 10918)"
+    'dc:creator': 'Michael Hanke',
+    'dc:description': 'dlsubject',
+    'dc:description<?xml:lang>': 'x-default',
+    'dc:title': 'dltitle',
+    'dc:title<?xml:lang>': 'x-default',
+    'pdfaid:part': '1',
+    'pdfaid:conformance': 'A',
+    'pdf:Keywords': 'dlkeyword1 dlkeyword2',
+    'pdf:Producer': 'LibreOffice 5.2',
+    'xmp:CreateDate': '2017-10-08T10:27:06+02:00',
+    'xmp:CreatorTool': 'Writer',
 }
 
 
 @with_tempfile(mkdir=True)
-def test_image(path):
+def test_xmp(path):
     ds = Dataset(path).create()
-    ds.config.add('datalad.metadata.nativetype', 'image', where='dataset')
+    ds.config.add('datalad.metadata.nativetype', 'xmp', where='dataset')
     copy(
-        opj(dirname(dirname(dirname(__file__))), 'tests', 'data', 'exif.jpg'),
+        opj(dirname(dirname(dirname(__file__))), 'tests', 'data', 'xmp.pdf'),
         path)
     ds.add('.')
     ok_clean_git(ds.path)
     res = ds.aggregate_metadata()
     assert_status('ok', res)
-    res = ds.metadata('exif.jpg')
+    res = ds.metadata('xmp.pdf')
     assert_result_count(res, 1)
 
-    # from this parser
-    meta = res[0]['metadata']['image']
+    # from this extractor
+    meta = res[0]['metadata']['xmp']
     for k, v in target.items():
         eq_(meta[k], v)
